refactor: rename headers slise to rest slice
