+++ store/store.ts
@@ -1,7 +1,7 @@
 import { configureStore } from '@reduxjs/toolkit';
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
 
-import headersReducer from './header-slice';
+import headersReducer from './rest-slice';
 
 export const store = configureStore({
   reducer: { headers: headersReducer },
